+++ neo3-gui/neo3-gui.tests/Storage/TrackDB_Test.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Neo;
 using Neo.Common.Storage;
+using Neo.Common.Storage.SQLiteModules;
 using Neo.Extensions;
 using Neo.Models;
 using Neo.SmartContract.Native;
@@ -23,6 +24,17 @@ namespace neo3_gui.tests.Storage
         [Ignore("Requires TrackDB runtime initialization - integration test")]
         public async Task AddTransfer_Test()
         {
+            _db.CreateContract(new ContractEntity
+            {
+                Hash = NativeContract.NEO.Hash.ToBigEndianHex(),
+                Name = "NEO",
+                Symbol = "NEO",
+                Decimals = NativeContract.NEO.Decimals,
+                AssetType = AssetType.Nep17,
+                CreateTime = DateTime.UtcNow,
+                CreateTxId = UInt256.Zero.ToBigEndianHex(),
+            });
+
             var transfer = new TransferInfo()
             {
                 BlockHeight = 0,
@@ -31,7 +43,7 @@ namespace neo3_gui.tests.Storage
                     
                 Asset = NativeContract.NEO.Hash,
                 Amount = 1,
-                TxId = UInt256.Parse("0x0000000000000000000000000000000000000000000000000000000000000000"),
+                TxId = UInt256.Zero,
                 TimeStamp = DateTime.Now.ToTimestamp(),
                 //AssetInfo = new AssetInfo()
                 //{