+++ neo3-gui/neo3-gui/publish.sh
@@ -1,3 +1,4 @@
+node scripts/fetch-leveldbstore.js
 dotnet publish -c Release -o ClientApp/build-neo-node
 cd ClientApp
 npm install